requirements.txt
# --- Core ---
python-dotenv
boto3
s3fs                  

# --- LlamaIndex core ---
llama-index
llama-index-core

# --- Docling integration ---
docling
llama-index-readers-docling

# --- Pinecone integration ---
pinecone-client
llama-index-vector-stores-pinecone

# --- Embeddings / OpenAI ---
openai
llama-index-embeddings-openai

# --- Aurora ---
psycopg2-binary